requirements.txt
# Core dependencies
torch>=2.0.1
transformers>=4.30.0
vllm>=0.2.0
sentence-transformers>=2.2.0

# HuggingFace Evaluate ecosystem
evaluate>=0.4.0
datasets>=2.14.0

# Evaluation metrics
sacrebleu>=2.5.0
rouge-score>=0.1.2
bert-score>=0.3.13

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning utilities
scikit-learn>=1.3.0
faiss-cpu>=1.7.0

# Visualization and plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Configuration and utilities
pyyaml>=6.0
tqdm>=4.65.0
requests>=2.31.0

# Logging and monitoring
loguru>=0.7.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Optional dependencies for enhanced performance
# Uncomment if available:
triton-windows>=3.6.0
# torch-audio>=2.0.0
# torch-vision>=0.15.0
# apex>=0.1.0
# flash-attn>=2.3.0
# hf_xet>=0.3.0  # For improved Xet Storage performance
